Added new formatter options, fix vulnerabilities, remove support for Aura Helper CLI and added support to Aura Helper SF Plugin

+++ src/scripts/commands/scan.ts
@@ -1,6 +1,7 @@
 import { CoreUtils, FileChecker, PathUtils, ProcessFactory, ProcessHandler } from '@aurahelper/core';
 import { StrUtils } from '@aurahelper/core/dist/utils';
 import * as vscode from 'vscode';
+import { applicationContext } from '../core/applicationContext';
 import { Config } from '../core/config';
 import { Paths } from '../core/paths';
 import { DiagnosticsManager, NotificationManager } from '../output';
@@ -38,7 +39,7 @@ export function run(fileUri: vscode.Uri): void {
             filePath = fileUri.fsPath;
         } else {
             let editor = vscode.window.activeTextEditor;
-            if (editor){
+            if (editor) {
                 filePath = editor.document.uri.fsPath;
             }
         }
@@ -60,7 +61,7 @@ function scan(filePath: string) {
                 location: vscode.ProgressLocation.Notification,
                 title: 'Installing SFDX Scanner',
                 cancellable: false
-            }, async () => { 
+            }, async () => {
                 try {
                     const process = ProcessFactory.installUpdateSFDXScanner();
                     const response = await ProcessHandler.runProcess(process) as ScannerResponse;
@@ -68,7 +69,7 @@ function scan(filePath: string) {
                 } catch (error) {
                     NotificationManager.showInfo(`SFDX Scanner is now installed. Try to run scan command again`);
                 }
-                
+
             });
         });
     } else {
@@ -77,7 +78,7 @@ function scan(filePath: string) {
                 location: vscode.ProgressLocation.Notification,
                 title: "Installing SFDX Scanner",
                 cancellable: false
-            }, async () => { 
+            }, async () => {
                 const process = ProcessFactory.installUpdateSFDXScanner();
                 const response = await ProcessHandler.runProcess(process) as ScannerResponse;
                 NotificationManager.showInfo(`SFDX Scanner is now installed. Try to run scan command again`);
@@ -86,28 +87,28 @@ function scan(filePath: string) {
     }
 }
 
-function runScanner(filePath: string, installCallback: any){
+function runScanner(filePath: string, installCallback: any) {
     vscode.window.withProgress({
-		location: vscode.ProgressLocation.Notification,
-		title: "Analizing File " + filePath,
-		cancellable: false
-	}, async () => { 
+        location: vscode.ProgressLocation.Notification,
+        title: "Analizing File " + filePath,
+        cancellable: false
+    }, async () => {
         let pdmRuleSetFile;
         let esLintRuleSetFile;
-        if(Config.getConfig().metadata.scanPmdRuleSet){
+        if (Config.getConfig().metadata.scanPmdRuleSet) {
             try {
                 pdmRuleSetFile = PathUtils.getAbsolutePath(Config.getConfig().metadata.scanPmdRuleSet);
-                if(!FileChecker.isExists(pdmRuleSetFile)){
+                if (!FileChecker.isExists(pdmRuleSetFile)) {
                     NotificationManager.showError(`The PDM Rules file ${pdmRuleSetFile} does not exists`);
                 }
             } catch (error) {
                 NotificationManager.showError(`Wrong path ${pdmRuleSetFile} to the PDM rule set`);
             }
         }
-        if(Config.getConfig().metadata.scanEsLintRuleSet){
+        if (Config.getConfig().metadata.scanEsLintRuleSet) {
             try {
                 esLintRuleSetFile = PathUtils.getAbsolutePath(Config.getConfig().metadata.scanEsLintRuleSet);
-                if(!FileChecker.isExists(esLintRuleSetFile)){
+                if (!FileChecker.isExists(esLintRuleSetFile)) {
                     NotificationManager.showError(`The PDM Rules file ${esLintRuleSetFile} does not exists`);
                 }
             } catch (error) {
@@ -115,19 +116,19 @@ function runScanner(filePath: string, installCallback: any){
             }
         }
         try {
-            const process = ProcessFactory.runScanner(filePath, Config.getConfig().metadata.scanCategories);
-            const response = await ProcessHandler.runProcess(process) as ScannerResponse;   
-            if(response.status === 0){
+            const process = applicationContext.ahSFPluginInstalled ? ProcessFactory.runSFScanner(filePath, Config.getConfig().metadata.scanCategories) : ProcessFactory.runSFDXScanner(filePath, Config.getConfig().metadata.scanCategories);
+            const response = await ProcessHandler.runProcess(process) as ScannerResponse;
+            if (response.status === 0) {
                 console.log(response);
-                if(response.result.length > 0){
+                if (response.result.length > 0) {
                     let totalProbles = 0;
-                    for(const fileProblem of response.result){
+                    for (const fileProblem of response.result) {
                         let diags = [];
                         let path: vscode.Uri | undefined;
                         path = Paths.toURI(fileProblem.fileName);
                         const fileName = PathUtils.getBasename(fileProblem.fileName);
-                        for(const problem of fileProblem.violations){
-                            const endColumn = problem.line !== problem.endLine ? problem.line + 5: problem.endColumn;
+                        for (const problem of fileProblem.violations) {
+                            const endColumn = problem.line !== problem.endLine ? problem.line + 5 : problem.endColumn;
                             const range = new vscode.Range(parseInt(problem.line) - 1, parseInt(problem.column) - 1, parseInt(problem.line) - 1, parseInt(endColumn));
                             const diagnostic = new vscode.Diagnostic(range, `${problem.message}Category: ${problem.category}\nViolation: ${problem.ruleName}: ${problem.url}`, vscode.DiagnosticSeverity.Warning);
                             diagnostic.source = 'SFDX Scanner';
@@ -148,15 +149,16 @@ function runScanner(filePath: string, installCallback: any){
                 NotificationManager.showError(`An error ocurred while scaning ${filePath}.\n${response.message}`);
             }
         } catch (error) {
-            if(CoreUtils.Utils.isString(error) && StrUtils.containsIgnorecase(error as string, 'scanner:run is not a sfdx command')){
-                
-                NotificationManager.showConfirmDialog('SFDX Scanner is not installed. Do you want to install now?', async () => {
-                    if(installCallback){
-                        installCallback();
-                    }
-                }, () => {
-                    NotificationManager.showWarning(`"AuraHelper: Scan" command will not work until install SFDX Scanner`);
-                });
+            if (CoreUtils.Utils.isString(error) && StrUtils.containsIgnorecase(error as string, 'scanner:run is not a sfdx command')) {
+                if(applicationContext.ahSFDXPluginInstalled){
+                    NotificationManager.showConfirmDialog('SFDX Scanner is not installed. Do you want to install now?', async () => {
+                        if (installCallback) {
+                            installCallback();
+                        }
+                    }, () => {
+                        NotificationManager.showWarning(`"AuraHelper: Scan" command will not work until install SFDX Scanner`);
+                    });
+                }
             } else {
                 NotificationManager.showError(`An error ocurred while scaning ${filePath}.\n${error}`);
             }
